Refactoring app into several files. Moving icons out of navbar
